backend-render/requirements.txt
Flask
flask-cors
python-dotenv
PyPDF2
python-docx
docx2txt
python-pptx
google-generativeai
supabase
PyMuPDF
gunicorn
